add history

+++ src/app/services/functions.service.ts
@@ -9,7 +9,7 @@ export class FunctionsService {
 
   public formateDate(date: Date) {
     const day = date.getDate().toString().padStart(2, '0');
-    const month = date.toLocaleString('en-GB', { month: 'long' });
+    const month = date.toLocaleString('en-GB', { month: 'short' });
     const year = date.getFullYear();
     const hours = date.getHours().toString().padStart(2, '0')
     const minutes = date.getMinutes().toString().padStart(2, '0')
